+++ lib/distribute_reads.rb
@@ -32,5 +32,5 @@ Object.send :include, DistributeReads::GlobalMethods
 
 ActiveSupport.on_load(:active_job) do
   require "distribute_reads/job_methods"
-  extend DistributeReads::JobMethods
+  include DistributeReads::JobMethods
 end