+++ lib/distribute_reads/job_methods.rb
@@ -1,8 +1,20 @@
+require "active_support/concern"
+
 module DistributeReads
   module JobMethods
-    def distribute_reads(max_lag: nil)
-      around_perform do |job, block|
-        distribute_reads(max_lag: max_lag) { block.call }
+    extend ActiveSupport::Concern
+
+    included do
+      before_perform do
+        Makara::Context.set_current(Makara::Context.generate) unless DistributeReads.default_to_primary
+      end
+    end
+
+    class_methods do
+      def distribute_reads(*args)
+        around_perform do |job, block|
+          distribute_reads(*args) { block.call }
+        end
       end
     end
   end